Added filtering by board name, priority
