+++ src/main/java/org/app/controllers/TodoController.java
@@ -1,13 +1,19 @@
 package org.app.controllers;
 
+import java.time.LocalDate;
 import java.util.List;
+
+import org.app.exceptions.TaskNotFoundException;
 import org.app.model.TodoTask;
+import org.app.model.enums.Priority;
 import org.app.model.enums.Status;
 import org.app.service.TodoTaskService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
@@ -23,5 +29,16 @@ public class TodoController {
         model.addAttribute("selectedStatuses", statuses);
         model.addAttribute("statuses", Status.values());
         return "todos";
+    }
+
+    @GetMapping("/search")
+    public String searchTasks(Model model, String keyword) throws TaskNotFoundException {
+
+        if(keyword != null){
+            model.addAttribute("userBoards", todoService.findByKeyword(keyword));
+        }else{
+            model.addAttribute("userBoards", todoService.getAllTasks());
+        }
+        return "home :: #boardList";
     }
 }
