+++ src/main/resources/templates/fragments/addboard-fragment.html
@@ -29,9 +29,6 @@
         </div>
     </div>
 </div>
-<script type="text/javascript" th:src="@{/js/script.js}"></script>
-<script>
 
-</script>
 </body>
 </html>
