+++ src/main/resources/templates/fragments/header-fragment.html
@@ -6,9 +6,11 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons+Outlined"  rel="stylesheet">
   <link rel="stylesheet" th:href="@{/css/style.css}">
   <link rel="stylesheet" th:href="@{/css/modal.css}">
-  <link rel="stylesheet" type="text/css" href="navbar.css" th:href="@{/css/navbar.css}" media="screen">
+  <link rel="stylesheet" th:href="@{/css/navbar.css}">
   <link rel="stylesheet" type="text/css" href="sidebar.css" th:href="@{/css/sidebar.css}" media="screen">
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
+  <script type="text/javascript" th:src="@{/js/script.js}"></script>
+  <script type="text/javascript" th:src="@{/js/nav.js}"></script>
   <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
