+++ src/main/resources/templates/fragments/sidebar-fragment.html
@@ -18,15 +18,19 @@
                 <li>Goals</li>
             </ul>
             <a sec:authorize="isAuthenticated()">Work Space</a>
+            <ul sec:authorize="isAuthenticated()">
+                <li>
+                    <a>
+                        Add new project
+                    </a>
+                </li>
+                <li>Setting</li>
+            </ul>
             <a sec:authorize="isAuthenticated()">My page</a>
             <a th:href="@{/signup}" sec:authorize="isAnonymous()">
                 Sign up</a>
             <a th:href="@{/users/login}" sec:authorize="!isAuthenticated()">
                 Log in</a>
-            <form th:action="@{/users/logout}" method="post" sec:authorize="isAuthenticated()">
-                <input hidden type="submit" value="Log out"/>
-                <a th:href="@{/users/logout}">Log out</a>
-            </form>
 
             </div>
         </div>
